pyproject.toml
[project]
name = "chatbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
    "email-validator>=2.0.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
start = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "db"]
only-include = ["api", "db", "main.py"]
